code/backend/requirements.txt
pymongo>=4.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
requests
python-dotenv>=1.0.0
streamlit>=1.28.0

